capsem_python/pyproject.toml
[project]
name = "capsem"
version = "0.3.1"
description = "CAPSEM: Contextual Agent Privacy and Security Manager"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "litellm>=1.77.4",
    "networkx>=3.5",
    "pip>=25.2",
    "pydantic>=2.11.9",
    "python-dotenv>=1.1.1",
    "tabulate>=0.9.0",
]

[dependency-groups]
dev = [
    "google-adk>=1.14.1",
    "ipykernel>=6.30.1",
    "nbconvert>=7.16.6",
    "pyright>=1.1.405",
    "pytest>=8.4.2",
    "pytest-asyncio>=0.24.0",
    "litellm>=1.77.4",
    "google-genai>=1.38.0",
    "openai>=1.109.1",
]
adk = [
    "google-adk>=1.14.1",
    "litellm>=1.77.4",
]
pii = [
    "presidio-analyzer>=2.2.354",
]
all = [
    { include-group = "dev" },
    { include-group = "adk" },
    { include-group = "pii" },
]

[tool.setuptools.packages.find]
include = ["capsem*"]
